+++ common/src/nativeMain/kotlin/com/powersync/sync/UserAgent.native.kt
@@ -4,5 +4,4 @@ import com.powersync.build.LIBRARY_VERSION
 import kotlin.experimental.ExperimentalNativeApi
 
 @OptIn(ExperimentalNativeApi::class)
-internal actual fun userAgent(): String =
-    "PowerSync Kotlin SDK v$LIBRARY_VERSION (running on ${Platform.cpuArchitecture.name} ${Platform.osFamily.name})"
+internal actual fun userAgent(): String = "powersync-kotlin/v$LIBRARY_VERSION ${Platform.cpuArchitecture.name}/${Platform.osFamily.name}"