+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.9.0 (unreleased)
 
+- Updated user agent string formats to allow viewing version distributions in the new PowerSync dashboard.
 - Sync options: `newClientImplementation` is now the default.
 - Make `androidx.sqlite:sqlite-bundled` an API dependency of `:core` to avoid toolchain warnings.